test: trace execution and reduction of take/filter/compare/like on complex arrays
Promote the trace test harness module to a directory module
(test_harness/trace/mod.rs + trace/tests.rs) and add snapshot tests that
trace how the optimizer and executor handle common operations over
complex arrays:

- take (DictArray) over a chunked array: no reduce rule fires; execution
  canonicalizes the chunked values chunk-by-chunk before the dict take
  kernel runs.
- filter over a struct array with complex children (dict-of-strings and
  chunked fields): FilterStructRule pushes the filter into each field and
  FilterReduceAdaptor(Dict) rewrites the dict field at optimize time.
- compare (Binary Eq vs constant) over a dict array:
  DictionaryScalarFnValuesPushDownRule reduces the compare into the dict
  values; execution shows the values-side binary kernel and the dict
  decode kernel.
- like (vs constant pattern) over a dict-of-strings array:
  LikeReduceAdaptor(Dict) reduces the like into the dict values.

Also add equivalent trace tests for run-end arrays in vortex-runend
(with insta as a new dev-dependency): compare reduces via
RunEndScalarFnRule at optimize time, while filter and take execute via
the FilterExecuteAdaptor/TakeExecuteAdaptor parent kernels.
